+++ src/main/java/com/raoulvdberge/refinedpipes/render/ItemPipeBakedModel.java
@@ -21,8 +21,9 @@ public class ItemPipeBakedModel implements IBakedModel {
     private final IBakedModel extension;
     private final IBakedModel straight;
     private final Map<AttachmentType, Map<Direction, IBakedModel>> attachmentModels = new HashMap<>();
+    private final Map<Direction, IBakedModel> inventoryAttachmentModels = new HashMap<>();
 
-    public ItemPipeBakedModel(IBakedModel core, IBakedModel extension, IBakedModel straight, Map<AttachmentType, IBakedModel> attachmentModels) {
+    public ItemPipeBakedModel(IBakedModel core, IBakedModel extension, IBakedModel straight, IBakedModel inventoryAttachment, Map<AttachmentType, IBakedModel> attachmentModels) {
         this.core = core;
         this.extension = extension;
         this.straight = straight;
@@ -38,6 +39,13 @@ public class ItemPipeBakedModel implements IBakedModel {
             }
 
             this.attachmentModels.put(entry.getKey(), dirToModel);
+        }
+
+        for (Direction direction : Direction.values()) {
+            this.inventoryAttachmentModels.put(direction, new TrsrBakedModel(
+                inventoryAttachment,
+                direction
+            ));
         }
     }
 
@@ -70,6 +78,7 @@ public class ItemPipeBakedModel implements IBakedModel {
             } else {
                 quads.addAll(core.getQuads(state, side, rand, EmptyModelData.INSTANCE));
 
+                // TODO: cache extensions as well.
                 if (north) {
                     quads.addAll(extension.getQuads(state, side, rand, EmptyModelData.INSTANCE));
                 }
@@ -100,6 +109,39 @@ public class ItemPipeBakedModel implements IBakedModel {
         if (attachments != null) {
             for (Map.Entry<Direction, AttachmentType> entry : attachments.entrySet()) {
                 quads.addAll(attachmentModels.get(entry.getValue()).get(entry.getKey()).getQuads(state, side, rand, extraData));
+            }
+        }
+
+        if (state != null) {
+            boolean invNorth = state.get(ItemPipeBlock.INV_NORTH);
+            boolean invEast = state.get(ItemPipeBlock.INV_EAST);
+            boolean invSouth = state.get(ItemPipeBlock.INV_SOUTH);
+            boolean invWest = state.get(ItemPipeBlock.INV_WEST);
+            boolean invUp = state.get(ItemPipeBlock.INV_UP);
+            boolean invDown = state.get(ItemPipeBlock.INV_DOWN);
+
+            if (invNorth && (attachments == null || !attachments.containsKey(Direction.NORTH))) {
+                quads.addAll(inventoryAttachmentModels.get(Direction.NORTH).getQuads(state, side, rand, EmptyModelData.INSTANCE));
+            }
+
+            if (invEast && (attachments == null || !attachments.containsKey(Direction.EAST))) {
+                quads.addAll(inventoryAttachmentModels.get(Direction.EAST).getQuads(state, side, rand, EmptyModelData.INSTANCE));
+            }
+
+            if (invSouth && (attachments == null || !attachments.containsKey(Direction.SOUTH))) {
+                quads.addAll(inventoryAttachmentModels.get(Direction.SOUTH).getQuads(state, side, rand, EmptyModelData.INSTANCE));
+            }
+
+            if (invWest && (attachments == null || !attachments.containsKey(Direction.WEST))) {
+                quads.addAll(inventoryAttachmentModels.get(Direction.WEST).getQuads(state, side, rand, EmptyModelData.INSTANCE));
+            }
+
+            if (invUp && (attachments == null || !attachments.containsKey(Direction.UP))) {
+                quads.addAll(inventoryAttachmentModels.get(Direction.UP).getQuads(state, side, rand, EmptyModelData.INSTANCE));
+            }
+
+            if (invDown && (attachments == null || !attachments.containsKey(Direction.DOWN))) {
+                quads.addAll(inventoryAttachmentModels.get(Direction.DOWN).getQuads(state, side, rand, EmptyModelData.INSTANCE));
             }
         }