+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.1
 - Renamed "Pipe" to "Item Pipe" (raoulvdberge)
 - Added different tiers for the Extractor Attachment: Basic, Improved and Advanced (raoulvdberge)
+- Fixed missing model variant errors (raoulvdberge)
 
 ## 0.1
 - Initial release