+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Refined Pipes Changelog
 
 ## 0.1.1
+Note: Pipes that are placed in the world from version 0.1 will all be removed, since the IDs have changed.
+
 - Renamed "Pipe" to "Item Pipe" (raoulvdberge)
 - Added different tiers for the Extractor Attachment: Basic, Improved and Advanced (raoulvdberge)
 - Fixed missing model variant errors (raoulvdberge)
+- Fixed pipes having no drops (raoulvdberge)
 
 ## 0.1
 - Initial release